+++ include/ftxui/screen/screen.hpp
@@ -28,13 +28,15 @@ struct Pixel {
   bool dim : 1;
   bool inverted : 1;
   bool underlined : 1;
+  bool automerge : 1;
 
   Pixel()
       : blink(false),
         bold(false),
         dim(false),
         inverted(false),
-        underlined(false) {}
+        underlined(false),
+        automerge(false) {}
 };
 
 /// @brief Define how the Screen's dimensions should look like.
@@ -62,8 +64,8 @@ class Screen {
   void Print();
 
   // Get screen dimensions.
-  int dimx() { return dimx_; }
-  int dimy() { return dimy_; }
+  int dimx() const { return dimx_; }
+  int dimy() const { return dimy_; }
 
   // Move the terminal cursor n-lines up with n = dimy().
   std::string ResetPosition(bool clear = false);