+++ src/ftxui/component/screen_interactive.cpp
@@ -1,19 +1,23 @@
 #include <stdio.h>    // for fileno, stdin
 #include <algorithm>  // for copy, max, min
-#include <csignal>  // for signal, SIGABRT, SIGFPE, SIGILL, SIGINT, SIGSEGV, SIGTERM, SIGWINCH
-#include <cstdlib>           // for NULL
+#include <chrono>  // for operator-, duration, operator>=, milliseconds, time_point, common_type<>::type
+#include <csignal>  // for signal, raise, SIGTSTP, SIGABRT, SIGFPE, SIGILL, SIGINT, SIGSEGV, SIGTERM, SIGWINCH
+#include <cstdlib>  // for NULL
+#include <functional>        // for function
 #include <initializer_list>  // for initializer_list
 #include <iostream>  // for cout, ostream, basic_ostream, operator<<, endl, flush
 #include <stack>     // for stack
-#include <thread>    // for thread
-#include <utility>   // for swap, move
-#include <vector>    // for vector
+#include <thread>    // for thread, sleep_for
+#include <type_traits>  // for decay_t
+#include <utility>      // for swap, move
+#include <variant>      // for visit
+#include <vector>       // for vector
 
+#include "ftxui/component/animation.hpp"  // for TimePoint, Clock, Duration, Params, RequestAnimationFrame
 #include "ftxui/component/captured_mouse.hpp"  // for CapturedMouse, CapturedMouseInterface
 #include "ftxui/component/component_base.hpp"  // for ComponentBase
 #include "ftxui/component/event.hpp"           // for Event
-#include "ftxui/component/mouse.hpp"           // for Mouse
-#include "ftxui/component/receiver.hpp"  // for ReceiverImpl, MakeReceiver, Sender, SenderImpl, Receiver
+#include "ftxui/component/receiver.hpp"  // for ReceiverImpl, Sender, MakeReceiver, SenderImpl, Receiver
 #include "ftxui/component/screen_interactive.hpp"
 #include "ftxui/component/terminal_input_parser.hpp"  // for TerminalInputParser
 #include "ftxui/dom/node.hpp"                         // for Node, Render
@@ -32,8 +36,8 @@
 #endif
 #else
 #include <sys/select.h>  // for select, FD_ISSET, FD_SET, FD_ZERO, fd_set
-#include <termios.h>     // for tcsetattr, tcgetattr, cc_t
-#include <unistd.h>      // for STDIN_FILENO, read
+#include <termios.h>  // for tcsetattr, termios, tcgetattr, TCSANOW, cc_t, ECHO, ICANON, VMIN, VTIME
+#include <unistd.h>  // for STDIN_FILENO, read
 #endif
 
 // Quick exit is missing in standard CLang headers
@@ -43,7 +47,17 @@
 
 namespace ftxui {
 
+namespace animation {
+void RequestAnimationFrame() {
+  auto* screen = ScreenInteractive::Active();
+  if (screen)
+    screen->RequestAnimationFrame();
+}
+}  // namespace animation
+
 namespace {
+
+ScreenInteractive* g_active_screen = nullptr;
 
 void Flush() {
   // Emscripten doesn't implement flush. We interpret zero as flush.
@@ -54,7 +68,7 @@ constexpr int timeout_milliseconds = 20;
 constexpr int timeout_microseconds = timeout_milliseconds * 1000;
 #if defined(_WIN32)
 
-void EventListener(std::atomic<bool>* quit, Sender<Event> out) {
+void EventListener(std::atomic<bool>* quit, Sender<Task> out) {
   auto console = GetStdHandle(STD_INPUT_HANDLE);
   auto parser = TerminalInputParser(out->Clone());
   while (!*quit) {
@@ -104,7 +118,7 @@ void EventListener(std::atomic<bool>* quit, Sender<Event> out) {
 #include <emscripten.h>
 
 // Read char from the terminal.
-void EventListener(std::atomic<bool>* quit, Sender<Event> out) {
+void EventListener(std::atomic<bool>* quit, Sender<Task> out) {
   (void)timeout_microseconds;
   auto parser = TerminalInputParser(std::move(out));
 
@@ -131,7 +145,7 @@ int CheckStdinReady(int usec_timeout) {
 }
 
 // Read char from the terminal.
-void EventListener(std::atomic<bool>* quit, Sender<Event> out) {
+void EventListener(std::atomic<bool>* quit, Sender<Task> out) {
   const int buffer_size = 100;
 
   auto parser = TerminalInputParser(std::move(out));
@@ -200,7 +214,7 @@ const std::string DeviceStatusReport(DSRMode ps) {
 }
 
 using SignalHandler = void(int);
-std::stack<ScreenInteractive::Callback> on_exit_functions;
+std::stack<Closure> on_exit_functions;
 void OnExit(int signal) {
   (void)signal;
   while (!on_exit_functions.empty()) {
@@ -211,12 +225,16 @@ void OnExit(int signal) {
 
 auto install_signal_handler = [](int sig, SignalHandler handler) {
   auto old_signal_handler = std::signal(sig, handler);
-  on_exit_functions.push([&] { std::signal(sig, old_signal_handler); });
+  on_exit_functions.push([=] { std::signal(sig, old_signal_handler); });
 };
 
-ScreenInteractive::Callback on_resize = [] {};
+Closure on_resize = [] {};
 void OnResize(int /* signal */) {
   on_resize();
+}
+
+void OnSigStop(int /*signal*/) {
+  ScreenInteractive::Private::SigStop(*g_active_screen);
 }
 
 class CapturedMouseImpl : public CapturedMouseInterface {
@@ -228,9 +246,14 @@ class CapturedMouseImpl : public CapturedMouseInterface {
   std::function<void(void)> callback_;
 };
 
-}  // namespace
+void AnimationListener(std::atomic<bool>* quit, Sender<Task> out) {
+  while (!*quit) {
+    out->Send(AnimationTask());
+    std::this_thread::sleep_for(std::chrono::milliseconds(15));
+  }
+}
 
-ScreenInteractive* g_active_screen = nullptr;
+}  // namespace
 
 ScreenInteractive::ScreenInteractive(int dimx,
                                      int dimy,
@@ -239,8 +262,7 @@ ScreenInteractive::ScreenInteractive(int dimx,
     : Screen(dimx, dimy),
       dimension_(dimension),
       use_alternative_screen_(use_alternative_screen) {
-  event_receiver_ = MakeReceiver<Event>();
-  event_sender_ = event_receiver_->MakeSender();
+  task_receiver_ = MakeReceiver<Task>();
 }
 
 // static
@@ -263,9 +285,21 @@ ScreenInteractive ScreenInteractive::FitComponent() {
   return ScreenInteractive(0, 0, Dimension::FitComponent, false);
 }
 
-void ScreenInteractive::PostEvent(Event event) {
+void ScreenInteractive::Post(Task task) {
   if (!quit_)
-    event_sender_->Send(event);
+    task_sender_->Send(task);
+}
+void ScreenInteractive::PostEvent(Event event) {
+  Post(event);
+}
+
+void ScreenInteractive::RequestAnimationFrame() {
+  if (animation_requested_)
+    return;
+  animation_requested_ = true;
+  auto now = animation::Clock::now();
+  if (now - previous_animation_time >= std::chrono::milliseconds(33))
+    previous_animation_time = now;
 }
 
 CapturedMouse ScreenInteractive::CaptureMouse() {
@@ -277,7 +311,6 @@ CapturedMouse ScreenInteractive::CaptureMouse() {
 }
 
 void ScreenInteractive::Loop(Component component) {
-
   // Suspend previously active screen:
   if (g_active_screen) {
     std::swap(suspended_screen_, g_active_screen);
@@ -315,12 +348,17 @@ void ScreenInteractive::Loop(Component component) {
 /// @brief Decorate a function. It executes the same way, but with the currently
 /// active screen terminal hooks temporarilly uninstalled during its execution.
 /// @param fn The function to decorate.
-ScreenInteractive::Callback ScreenInteractive::WithRestoredIO(Callback fn) {
+Closure ScreenInteractive::WithRestoredIO(Closure fn) {
   return [this, fn] {
     Uninstall();
     fn();
     Install();
   };
+}
+
+// static
+ScreenInteractive* ScreenInteractive::Active() {
+  return g_active_screen;
 }
 
 void ScreenInteractive::Install() {
@@ -382,8 +420,11 @@ void ScreenInteractive::Install() {
   tcsetattr(STDIN_FILENO, TCSANOW, &terminal);
 
   // Handle resize.
-  on_resize = [&] { event_sender_->Send(Event::Special({0})); };
+  on_resize = [&] { task_sender_->Send(Event::Special({0})); };
   install_signal_handler(SIGWINCH, OnResize);
+
+  // Handle SIGTSTP/SIGCONT.
+  install_signal_handler(SIGTSTP, OnSigStop);
 #endif
 
   auto enable = [&](std::vector<DECMode> parameters) {
@@ -419,43 +460,88 @@ void ScreenInteractive::Install() {
   Flush();
 
   quit_ = false;
+  task_sender_ = task_receiver_->MakeSender();
   event_listener_ =
-      std::thread(&EventListener, &quit_, event_receiver_->MakeSender());
+      std::thread(&EventListener, &quit_, task_receiver_->MakeSender());
+  animation_listener_ =
+      std::thread(&AnimationListener, &quit_, task_receiver_->MakeSender());
 }
 
 void ScreenInteractive::Uninstall() {
   ExitLoopClosure()();
   event_listener_.join();
+  animation_listener_.join();
 
   OnExit(0);
 }
 
 void ScreenInteractive::Main(Component component) {
+  previous_animation_time = animation::Clock::now();
+
+  auto draw = [&] {
+    Draw(component);
+    std::cout << ToString() << set_cursor_position;
+    Flush();
+    Clear();
+  };
+
+  draw();
+
   while (!quit_) {
-    if (!event_receiver_->HasPending()) {
-      Draw(component);
-      std::cout << ToString() << set_cursor_position;
-      Flush();
-      Clear();
+    if (!task_receiver_->HasPending())
+      draw();
+
+    bool continue_event_loop = true;
+    while (continue_event_loop) {
+      continue_event_loop = false;
+      Task task;
+      if (!task_receiver_->Receive(&task))
+        break;
+
+      std::visit(
+          [&](auto&& arg) {
+            using T = std::decay_t<decltype(arg)>;
+
+            // Handle Event.
+            if constexpr (std::is_same_v<T, Event>) {
+              if (arg.is_cursor_reporting()) {
+                cursor_x_ = arg.cursor_x();
+                cursor_y_ = arg.cursor_y();
+                return;
+              }
+
+              if (arg.is_mouse()) {
+                arg.mouse().x -= cursor_x_;
+                arg.mouse().y -= cursor_y_;
+              }
+
+              arg.screen_ = this;
+              component->OnEvent(arg);
+            }
+
+            // Handle callback
+            if constexpr (std::is_same_v<T, Closure>) {
+              arg();
+              return;
+            }
+
+            // Handle Animation
+            if constexpr (std::is_same_v<T, AnimationTask>) {
+              if (!animation_requested_) {
+                continue_event_loop = true;
+                return;
+              }
+              animation_requested_ = false;
+              animation::TimePoint now = animation::Clock::now();
+              animation::Duration delta = now - previous_animation_time;
+              previous_animation_time = now;
+
+              animation::Params params(delta);
+              component->OnAnimation(params);
+            }
+          },
+          task);
     }
-
-    Event event;
-    if (!event_receiver_->Receive(&event))
-      break;
-
-    if (event.is_cursor_reporting()) {
-      cursor_x_ = event.cursor_x();
-      cursor_y_ = event.cursor_y();
-      continue;
-    }
-
-    if (event.is_mouse()) {
-      event.mouse().x -= cursor_x_;
-      event.mouse().y -= cursor_y_;
-    }
-
-    event.screen_ = this;
-    component->OnEvent(event);
   }
 }
 
@@ -538,11 +624,29 @@ void ScreenInteractive::Draw(Component component) {
   }
 }
 
-ScreenInteractive::Callback ScreenInteractive::ExitLoopClosure() {
+Closure ScreenInteractive::ExitLoopClosure() {
   return [this] {
     quit_ = true;
-    event_sender_.reset();
+    task_sender_.reset();
   };
+}
+
+void ScreenInteractive::SigStop() {
+#if defined(_WIN32)
+  // Windows do no support SIGTSTP.
+#else
+  Post([&] {
+    Uninstall();
+    std::cout << reset_cursor_position;
+    reset_cursor_position = "";
+    std::cout << ResetPosition(/*clear=*/true);
+    dimx_ = 0;
+    dimy_ = 0;
+    Flush();
+    std::raise(SIGTSTP);
+    Install();
+  });
+#endif
 }
 
 }  // namespace ftxui.