+++ src/ftxui/component/terminal_input_parser.cpp
@@ -1,15 +1,15 @@
 #include "ftxui/component/terminal_input_parser.hpp"
 
-#include <algorithm>  // for max
-#include <cstdint>
+#include <cstdint>  // for uint32_t
 #include <memory>   // for unique_ptr
 #include <utility>  // for move
 
 #include "ftxui/component/event.hpp"  // for Event
+#include "ftxui/component/task.hpp"   // for Task
 
 namespace ftxui {
 
-TerminalInputParser::TerminalInputParser(Sender<Event> out)
+TerminalInputParser::TerminalInputParser(Sender<Task> out)
     : out_(std::move(out)) {}
 
 void TerminalInputParser::Timeout(int time) {
@@ -52,7 +52,14 @@ void TerminalInputParser::Send(TerminalInputParser::Output output) {
       return;
 
     case SPECIAL:
-      out_->Send(Event::Special(std::move(pending_)));
+      // Microsoft's terminal uses a different new line character for the return
+      // key. This also happens with linux with the `bind` command:
+      // See https://github.com/ArthurSonzogni/FTXUI/issues/337
+      // Here, we uniformize the new line character to `\n`.
+      if (pending_ == "\r")
+        out_->Send(Event::Special("\n"));
+      else
+        out_->Send(Event::Special(std::move(pending_)));
       pending_.clear();
       return;